added structure to the app.vue, added sass module, added styles directory with all nessesary files
