+++ src/components/Footer/Footer.vue
@@ -1,0 +1,5 @@
+<template>
+  <footer>
+    FOOTER
+  </footer>
+</template>
