+++ src/components/Header/Header.vue
@@ -1,0 +1,5 @@
+<template>
+  <header>
+    HEADER
+  </header>
+</template>
