+++ src/components/Steps/Steps.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="global">
+    Steps
+  </div>
+</template>
