+++ src/styles/_core.scss
@@ -1,0 +1,15 @@
+//
+//  Gene Vue Checkout App
+//  ---------------------------------------------
+
+
+//
+//  Core Checkout Styles
+//  ---------------------------------------------
+
+@import 'core/_loader';
+@import 'core/_form';
+@import 'core/_button';
+
+@import 'core/_layout';
+@import 'core/_steps';
